Add aggregated clusterroles for `view` and `cluster-reader`
Kyverno already ships with aggregated clusterroles for `admin`, but
doesn't provide read-only variants for `view` and `cluster-reader`.

We possibly could extract the kinds from the upstream aggregated roles
for admin, but they're not separated by cluster-scoped and namespaced
resources, so we'd have to do some manual checking anyway when updating
to a new minor version.